🚚 rename: 퀴즈리스트, 아이템에서 사용된 함수들 냅다 util로 이동
